+++ components/index.ts
@@ -27,7 +27,8 @@ export { Layout } from "./Layout/Layout";
 export { Loader } from "./Loader";
 export { LoadingSkeleton } from "./LoadingSkeleton/LoadingSkeleton";
 export { LoadingSpinner } from "./LoadingSpinner/LoadingSpinner";
-export { MigrationBanner } from "./MigrationBanner/MigrationBanner";
+export { MigrationBanner } from "./AnnouncementBanners/MigrationBanner";
+export { GasRebateBanner } from "./AnnouncementBanners/GasRebateBanner";
 export { Nav } from "./Nav/Nav";
 export { Notifications } from "./Notifications/Notifications";
 export { Pagination, usePagination } from "./Pagination/Pagination";
